+++ R/attr-nodes.R
@@ -71,12 +71,49 @@ digest_curly <- function(curly, ns) {
 #' m$body <- protect_curly(m$body)
 #' xml2::xml_child(m$body)
 protect_curly <- function(body, ns = md_ns()) {
-  body  <- copy_xml(body)
-  curly  <- find_curly(body, ns)
+  body <- copy_xml(body)
+  curly <- find_curly(body, ns)
   new_nodes <- purrr::map(curly, digest_curly, ns = ns)
   # since we split up the nodes, we have to do this node by node
   for (i in seq(new_nodes)) {
     add_node_siblings(curly[[i]], new_nodes[[i]], remove = TRUE)
   }
   copy_xml(body)
+}
+
+#' Protect fences of Pandoc fences divs for further processing
+#'
+#' @inheritParams protect_math
+#' @return a copy of the modified XML object
+#' @details Commonmark will render text such as `::: footer`
+#' as normal text which might be problematic if trying to extract
+#' real text from the XML.
+#'
+#' If sending the XML to, say, a translation API that allows some tags
+#' to be ignored, you could first transform the text tags with the
+#' attribute `fences` to `fences` tags, and then transform them back
+#' to text tags before using `to_md()`.
+#'
+#' @note this function is also a method in the [tinkr::yarn] object.
+#'
+#' @export
+#' @examples
+#' m <- tinkr::to_xml(system.file("extdata", "fenced-divs.md", package = "tinkr"))
+#' xml2::xml_child(m$body)
+#' m$body <- protect_fences(m$body)
+#' xml2::xml_child(m$body)
+protect_fences <- function(body, ns = md_ns()) {
+  body <- copy_xml(body)
+  fences <- find_fences(body, ns)
+  purrr::map(fences, digest_fence, ns = ns)
+  body
+}
+
+find_fences <- function(body, ns) {
+  i <- ".//md:text[starts-with(text(),':::')]"
+  xml2::xml_find_all(body, i, ns = ns)
+}
+
+digest_fence <- function(fence, ns) {
+  xml2::xml_attr(fence, "fence") <- "true"
 }