+++ R/to_md.R
@@ -1,6 +1,7 @@
-#' Write YAML and XML back to disk as (R)Markdown
+#' Write front-matter (YAML, TOML or JSON) and XML back to disk as (R)Markdown
 #'
-#' @param yaml_xml_list result from a call to [to_xml()] and editing.
+#' @param frontmatter_xml_list result from a call to [to_xml()] and editing.
+#' @param yaml_xml_list `r lifecycle::badge('deprecated')` Use `frontmatter_xml_list()`.
 #' @param path path of the new file. Defaults to `NULL`, which will not write
 #'   any file, but will still produce the conversion and pass the output as
 #'   a character vector.
@@ -13,20 +14,20 @@
 #' the path to your XSL stylesheet as argument.
 #'
 #'
-#' @return 
-#'  - `to_md()`: `\[character\]` the converted document, invisibly as a character vector containing two elements: the yaml list and the markdown body.
+#' @return
+#'  - `to_md()`: `\[character\]` the converted document, invisibly as a character vector containing two elements: the frontmatter list and the markdown body.
 #'  - `to_md_vec()`: `\[character\]` the markdown representation of each node.
 #'
 #' @export
 #'
 #' @examples
 #' path <- system.file("extdata", "example1.md", package = "tinkr")
-#' yaml_xml_list <- to_xml(path)
-#' names(yaml_xml_list)
+#' frontmatter_xml_list <- to_xml(path)
+#' names(frontmatter_xml_list)
 #' # extract the level 3 headers from the body
 #' headers3 <- xml2::xml_find_all(
-#'   yaml_xml_list$body,
-#'   xpath = './/md:heading[@level="3"]', 
+#'   frontmatter_xml_list$body,
+#'   xpath = './/md:heading[@level="3"]',
 #'   ns = md_ns()
 #' )
 #' # show the headers
@@ -39,17 +40,21 @@
 #' print(h1 <- to_md_vec(headers3))
 #' # save back and have a look
 #' newmd <- tempfile("newmd", fileext = ".md")
-#' res <- to_md(yaml_xml_list, newmd)
+#' res <- to_md(frontmatter_xml_list, newmd)
 #' # show that it works
 #' regmatches(res[[2]], gregexpr(h1[1], res[[2]], fixed = TRUE))
 #' # file.edit("newmd.md")
 #' file.remove(newmd)
-#' 
-to_md <- function(yaml_xml_list, path = NULL, stylesheet_path = stylesheet()){
-
+#'
+to_md <- function(
+  frontmatter_xml_list,
+  path = NULL,
+  stylesheet_path = stylesheet(),
+  yaml_xml_list = deprecated()
+) {
   # duplicate document to avoid overwriting
-  body <- copy_xml(yaml_xml_list$body)
-  yaml <- yaml_xml_list$yaml
+  body <- copy_xml(frontmatter_xml_list$body)
+  frontmatter <- frontmatter_xml_list$frontmatter
 
   # read stylesheet and fail early if it doesn't exist
   stylesheet <- read_stylesheet(stylesheet_path)
@@ -57,10 +62,10 @@ to_md <- function(yaml_xml_list, path = NULL, stylesheet_path = stylesheet()){
   transform_code_blocks(body)
   remove_phantom_text(body)
 
-  md_out <- transform_to_md(body, yaml, stylesheet)
+  md_out <- transform_to_md(body, frontmatter, stylesheet)
 
   if (!is.null(path)) {
-    writeLines(md_out, con = path, useBytes = TRUE, sep =  "\n\n")
+    writeLines(md_out, con = path, useBytes = TRUE, sep = "\n\n")
   }
 
   invisible(md_out)
@@ -80,15 +85,15 @@ to_md_vec <- function(nodelist, stylesheet_path = stylesheet()) {
 }
 
 
-# convert body and yaml to markdown text given a stylesheet
-transform_to_md <- function(body, yaml, stylesheet) {
+# convert body and frontmatter to markdown text given a stylesheet
+transform_to_md <- function(body, frontmatter, stylesheet) {
   body <- xslt::xml_xslt(body, stylesheet = stylesheet)
 
-  if (length(yaml) > 0) {
-    yaml <- glue::glue_collapse(yaml, sep = "\n")
+  if (length(frontmatter) > 0) {
+    frontmatter <- glue::glue_collapse(frontmatter, sep = "\n")
   }
 
-  c(yaml, body)
+  c(frontmatter, body)
 }
 
 # remove phantom text nodes that occur before links, images, and asis nodes that
@@ -99,8 +104,11 @@ remove_phantom_text <- function(body) {
   # to_protect <- xml2::xml_find_all(body,
   #   ".//md:link | .//md:image | .//md:text[@asis]", ns = md_ns())
   # # find the nodes that precede these nodes with zero length text
-  to_sever <- xml2::xml_find_all(body,
-    ".//md:text[string-length(text())=0]", ns = md_ns())
+  to_sever <- xml2::xml_find_all(
+    body,
+    ".//md:text[string-length(text())=0]",
+    ns = md_ns()
+  )
   if (length(to_sever)) {
     xml2::xml_remove(to_sever)
   }
@@ -111,13 +119,12 @@ copy_xml <- function(xml) {
   xml2::read_xml(as.character(xml))
 }
 
-transform_code_blocks <- function(xml){
+transform_code_blocks <- function(xml) {
   # Find all code blocks with a language attribute (those without it are not processed)
   code_blocks <- xml %>%
-    xml2::xml_find_all(xpath = './/d1:code_block[@language]',
-                       xml2::xml_ns(.))
+    xml2::xml_find_all(xpath = './/d1:code_block[@language]', xml2::xml_ns(.))
 
-  if(length(code_blocks) == 0){
+  if (length(code_blocks) == 0) {
     return(TRUE)
   }
 
@@ -126,28 +133,29 @@ transform_code_blocks <- function(xml){
   purrr::walk(code_blocks, to_info)
 }
 
-to_info <- function(code_block){
- attrs <- xml2::xml_attrs(code_block)
- options <- attrs[!names(attrs) %in%
-                  c("language", "name", "space", "sourcepos", "xmlns", "xmlns:xml")]
+to_info <- function(code_block) {
+  attrs <- xml2::xml_attrs(code_block)
+  options <- attrs[
+    !names(attrs) %in%
+      c("language", "name", "space", "sourcepos", "xmlns", "xmlns:xml")
+  ]
 
- if(length(options) > 0){
-   options <- glue::glue("{names(options)}={options}") %>%
-     glue::glue_collapse(sep = ", ")
-   options <- paste(",", options)
- }else{
-   options <- ""
- }
+  if (length(options) > 0) {
+    options <- glue::glue("{names(options)}={options}") %>%
+      glue::glue_collapse(sep = ", ")
+    options <- paste(",", options)
+  } else {
+    options <- ""
+  }
 
- if (attrs["name"] != ""){
-   attrs["name"] <- paste0(" ", attrs["name"])
- }
+  if (attrs["name"] != "") {
+    attrs["name"] <- paste0(" ", attrs["name"])
+  }
 
- info <- glue::glue('{attrs["language"]}{attrs["name"]}{options}')
- info <- paste0("{", info)
- info <- paste0(info, "}")
- names(info) <- "info"
+  info <- glue::glue('{attrs["language"]}{attrs["name"]}{options}')
+  info <- paste0("{", info)
+  info <- paste0(info, "}")
+  names(info) <- "info"
 
- xml2::xml_set_attr(code_block, "info", info)
+  xml2::xml_set_attr(code_block, "info", info)
 }
-