+++ R/to_xml.R
@@ -15,10 +15,11 @@
 #'   possible. If this is `FALSE`, these braces will be escaped in the output
 #'   document. See [protect_unescaped()] for details.
 #'
-#' @return A list containing the YAML of the file (yaml)
+#' @return A list containing the front-matter (YAML, TOML or JSON)
+#'   of the file (frontmatter)
 #'   and its body (body) as XML.
 #'
-#' @details This function will take a (R)markdown file, split the yaml header
+#' @details This function will take a (R)markdown file, split the frontmatter
 #'   from the body, and read in the body through [commonmark::markdown_xml()].
 #'   Any RMarkdown code fences will be parsed to expose the chunk options in
 #'   XML and tickboxes (aka checkboxes) in GitHub-flavored markdown will be
@@ -37,12 +38,18 @@
 #' path2 <- system.file("extdata", "example2.Rmd", package = "tinkr")
 #' post_list2 <- to_xml(path2)
 #' post_list2
-to_xml <- function(path, encoding = "UTF-8", sourcepos = FALSE, anchor_links = TRUE, unescaped = TRUE){
+to_xml <- function(
+  path,
+  encoding = "UTF-8",
+  sourcepos = FALSE,
+  anchor_links = TRUE,
+  unescaped = TRUE
+) {
   content <- readLines(path, encoding = encoding, warn = FALSE)
 
-  splitted_content <- split_yaml_body(content)
+  splitted_content <- split_frontmatter_body(content)
 
-  yaml <- splitted_content$yaml
+  frontmatter <- splitted_content$frontmatter
 
   splitted_content$body %>%
     clean_content() %>%
@@ -60,12 +67,15 @@ to_xml <- function(path, encoding = "UTF-8", sourcepos = FALSE, anchor_links = T
     body <- resolve_anchor_links(body, splitted_content$body)
   }
 
-  list(yaml = yaml,
-       body = body)
+  list(
+    frontmatter = frontmatter,
+    body = body,
+    frontmatter_format = splitted_content$frontmatter_format
+  )
 }
 
 
-clean_content <- function(content){
+clean_content <- function(content) {
   illegal_control_chars <- "[^\u0009\u000a\u000d\u0020-\uD7FF\uE000-\uFFFD]"
   smart_double_quotes <- "[\u201C\u201D]"
   smart_single_quotes <- "[\u2018\u2019]"
@@ -76,7 +86,7 @@ clean_content <- function(content){
 }
 
 
-transform_block <- function(code_block){
+transform_block <- function(code_block) {
   info <- xml2::xml_attr(code_block, "info")
 
   if (is.na(info) || !str_detect(info, "^\\{.+?\\}$")) {
@@ -100,10 +110,9 @@ transform_block <- function(code_block){
   code_block
 }
 
-parse_rmd <- function(body){
+parse_rmd <- function(body) {
   code_blocks <- body %>%
-    xml2::xml_find_all(xpath = './/d1:code_block',
-                       xml2::xml_ns(.))
+    xml2::xml_find_all(xpath = './/d1:code_block', xml2::xml_ns(.))
 
   purrr::walk(code_blocks, transform_block)
 }